create a better application structure used blueprint and create more sub directories for authentication, errors and main application code generate a requirements.txt file now test files can use their own configuration
