+++ .gitignore
@@ -4,3 +4,4 @@
 .DS_Store
 .idea/
 app.db
+.env
