+++ app/models.py
@@ -3,8 +3,9 @@ from datetime import datetime
 from time import time
 
 import jwt
+from flask import current_app
 from werkzeug.security import generate_password_hash, check_password_hash
-from app import db, login, app
+from app import db, login
 from flask_login import UserMixin
 
 
@@ -71,12 +72,12 @@ class User(UserMixin, db.Model):
     def get_reset_password_token(self, expires_in=600):
         return jwt.encode(
             {'reset_password': self.id, 'exp': time() + expires_in},
-            app.config['SECRET_KEY'], algorithm='HS256').decode('utf-8')
+            current_app.config['SECRET_KEY'], algorithm='HS256').decode('utf-8')
 
     @staticmethod
     def verify_reset_password_token(token):
         try:
-            id = jwt.decode(token, app.config['SECRET_KEY'],
+            id = jwt.decode(token, current_app.config['SECRET_KEY'],
                             algorithms=['HS256'])['reset_password']
         except:
             return
