+++ app/templates/_post.html
@@ -1,12 +1,12 @@
 <table class="table table-hover">
         <tr>
             <td width="70px">
-               <a href="{{ url_for('user', username=post.author.username) }}">
+               <a href="{{ url_for('main.user', username=post.author.username) }}">
                     <img src="{{ post.author.avatar(70) }}" />
                </a>
             </td>
             <td>
-                <a href="{{ url_for('user', username=post.author.username) }}">
+                <a href="{{ url_for('main.user', username=post.author.username) }}">
                     {{ post.author.username }}
                 </a>
                 said {{ moment(post.timestamp).fromNow() }}:
