+++ app/templates/email/reset_password.html
@@ -1,12 +1,12 @@
 <p>Dear {{ user.username }},</p>
 <p>
     To reset your password
-    <a href="{{ url_for('reset_password', token=token, _external=True) }}">
+    <a href="{{ url_for('auth.reset_password', token=token, _external=True) }}">
         click here
     </a>.
 </p>
 <p>Alternatively, you can paste the following link in your browser's address bar:</p>
-<p>{{ url_for('reset_password', token=token, _external=True) }}</p>
+<p>{{ url_for('auth.reset_password', token=token, _external=True) }}</p>
 <p>If you have not requested a password reset simply ignore this message.</p>
 <p>Sincerely,</p>
 <p>The Microblog Team</p>
