+++ app/templates/errors/404.html
@@ -2,5 +2,5 @@
 
 {% block app_content %}
     <h1>File Not Found</h1>
-    <p><a href="{{ url_for('index') }}">Back</a></p>
+    <p><a href="{{ url_for('main.index') }}">Back</a></p>
 {% endblock %}
