+++ microblog.py
@@ -1,5 +1,7 @@
-from app import app, db
+from app import db, create_app
 from app.models import User, Post
+
+app = create_app()
 
 
 @app.shell_context_processor
